+++ static/scripts/administration/classes/manage.js
@@ -2,33 +2,6 @@ import printQRs from '../../helpers/printQRs';
 
 /* globals populateModalForm */
 
-function sortOptions(selector, sortFunction) {
-	const input = document.querySelector(selector);
-	const options = Array.from(input.querySelectorAll('option'));
-
-	options.sort(sortFunction);
-	options.forEach((option) => {
-		input.appendChild(option);
-	});
-}
-
-function sortStudents() {
-	const sortFunction = (a, b) => {
-		const aValue = a.dataset.lastName;
-		const bValue = b.dataset.lastName;
-		if (aValue < bValue) {
-			return -1;
-		}
-		if (aValue > bValue) {
-			return 1;
-		}
-		return 0;
-	};
-
-	const studentInputSelector = 'select[name=userIds]';
-	sortOptions(studentInputSelector, sortFunction);
-}
-
 function copy(event) {
 	event.preventDefault();
 	const { copySelector } = event.target.dataset;
@@ -131,7 +104,6 @@ window.addEventListener('load', () => {
 				`option[value="${student._id}"]`,
 			).selected = true;
 		});
-		sortStudents();
 	});
 
 	function btnSendLinksEmailsHandler(e) {
@@ -222,5 +194,3 @@ window.addEventListener('load', () => {
 	}
 	$('.btn-print-links').on('click', btnPrintLinksHandler);
 });
-
-window.addEventListener('load', sortStudents);